Implement analogReadResolution
The current released version, was always reading with 12 bits of resolution.
The Arduino analogRead web page said that by default it should default to 10 bits.

The updated code is now ignoring the arduino_adc[idx].resolution
field as this is setup for 12 and you can not change it run time.

Like the MBED version,  this version is using the static variable read_resolution to hold the current resolution.  Which we then pass through to zephyr when we do the reads.

I also changed the parameter buf that we pass through to zephyr from in16_t to uint16_t as if you choose resolution of 16 bits it was returning negative values.