Portenta H7: extended pin list, pinNames, SPI, WIre, PWM, ADC
This is a replacement for PR: #71 and #82.

All of the earlier commits were squashed into one.  Then this was converted a few times during the #85 pr time
frame as things kept changing and moving around.   It has now been updated to the released .3 version.

I started off adding in the whole pin table as defined by the MBED version, which actually contained duplicate defines.  I later reduced
this set such that it now longer matches the MBED version, but does still include all of the pins that have external pins on some of the
breakout boards. As for compatibility, most of the documentation for these show the PIN names and not numbers, so I imported the
MBED Pin name table and have the start of allowing several different operations to be done, like pinMode, digitalWrite.

We defined the additional SPI ports and Wire ports.

We defined an initial setup for Analog pins.  Have similar hack to GIGA version for pure Analog.  Added additional hacks for duplicated
pins.  That is two of the analog Pins are the exact same pin as some other digital pins...

Added some PWM support.

Also added WIP: camera support.

Co-Authored-By: Mike S <5366213+mjs513@users.noreply.github.com>