GiGA Touch - Allow us to define a callback function and add a forwarder in fixups.
Added the ability to register a callback function to be called by the zephyr input
system.

This new code is only there if it is a GIGA and CONFIG_INPUT_GT911_INTERRUPT is defined.

This includes adding a callback function that is linked in to the zephyr build,  and export
a function to call to allow us to register our own.

Added an init variant to GIGA that if the touch interrupt is enabled, then it will clear out the old user defined
callback if any.

This version also added in the changes to fixups that allow a camera to work on Portenta H7.
Needed to have it do the same as GIGA to start the PWM if INPUT is configured in.

And updated the portentas remote_endpoint define to be using the same format as used in the giga overlay
Also changed to STM32_DMA_FIFO_1-4, like the GIGA currently has

Update arduino_portenta_h7_stm32h747xx_m7.overlay

Fix build

Fix the overlay, newer zephyr wants different form...