camera: try to defer init of video objects
This change, removes the automatic starting of the PWM clock on the
GIGA, at startup.  Instead it starts the clock if/when the sketch calls
the Camera::begin method.

But to make this work, we also need to not start up the video objects, until
after the MCLK has been started.  We can do that with marking them
as zephyr,deferred-init